+++ context/app/static/js/shared-styles/sections/TableOfContents/types.ts
@@ -1,0 +1,14 @@
+import SvgIcon from '@mui/material/SvgIcon';
+
+export interface TableOfContentsItem {
+  text: string;
+  hash: string;
+  icon?: typeof SvgIcon;
+  items?: TableOfContentsItem[];
+}
+
+export interface TableOfContentsItemWithNode extends TableOfContentsItem {
+  node: ReturnType<typeof document.getElementById>;
+}
+
+export type TableOfContentsItems<I = TableOfContentsItem> = I[];
